new icon created supports all res Fixed common hashing Fixed user prefix encryption Light Loc used on lookup network Search results now returned immediately un-reffed files in versioned cache and storage removed
git-svn-id: https://jmg.devguard.com/svn/deops@150 a8feb433-1630-0410-ae02-e0276e0da4ce

+++ Update/MainForm.cs
@@ -7,6 +7,7 @@ using System.Drawing;
 using System.IO;
 using System.Linq;
 using System.Text;
+using System.Threading;
 using System.Windows.Forms;
 
 namespace UpdateOp
@@ -14,6 +15,10 @@ namespace UpdateOp
     public partial class MainForm : Form
     {
         string NewExe;
+
+        Thread Worker;
+        bool KillThread;
+        AutoResetEvent TryUpdate = new AutoResetEvent(false);
 
         public MainForm(string newExe)
         {
@@ -23,28 +28,70 @@ namespace UpdateOp
                 BackColor = System.Drawing.Color.WhiteSmoke;
 
             NewExe = newExe;
+
+            Worker = new Thread(ApplyUpdate);
+            Worker.Start();
         }
 
         private void MainForm_Load(object sender, EventArgs e)
         {
-            ApplyUpdate();
+           TryUpdate.Set();
         }
 
-        void ApplyUpdate()
+        private void SecondTimer_Tick(object sender, EventArgs e)
         {
-            try
-            {
-                File.Copy(NewExe, "RiseOp.exe", true);
-                File.Delete(NewExe);
+            TryUpdate.Set();
+        }
 
-                Process.Start("RiseOp.exe");
-
-                Close();
-            }
-            catch (Exception ex)
+        public void ApplyUpdate()
+        {
+            while (!KillThread)
             {
-                StatusLabel.Text = "Could not apply update. Ensure RiseOp is closed.\n" + ex.Message;
+                TryUpdate.WaitOne();
+
+                if (KillThread)
+                    break;
+
+                try
+                {
+                    string windir = Environment.GetEnvironmentVariable("windir");
+                    string ngen = windir + "\\Microsoft.NET\\Framework\\v2.0.50727\\ngen.exe";
+
+
+                    RunInGui(() => StatusLabel.Text = "Uninstalling...");
+                    StartSilent(ngen, "uninstall RiseOp.exe");
+
+
+                    RunInGui(() => StatusLabel.Text = "Installing...");
+                    File.Copy(NewExe, "RiseOp.exe", true);
+                    File.Delete(NewExe);
+
+
+                    RunInGui(() => StatusLabel.Text = "Optimizing...");
+                    StartSilent(ngen, "install RiseOp.exe /queue");
+
+                    Process.Start("RiseOp.exe");
+
+                    RunInGui(() => Close());
+                }
+                catch (Exception ex)
+                {
+                    RunInGui(() => StatusLabel.Text = "Could not apply update. Ensure RiseOp is closed.\n" + ex.Message);
+                }
             }
+        }
+
+        private void StartSilent(string file, string args)
+        {
+            ProcessStartInfo info = new ProcessStartInfo(file, args);
+            info.WindowStyle = ProcessWindowStyle.Hidden;
+
+            Process.Start(info).WaitForExit();
+        }
+
+        void RunInGui(Action action)
+        {
+            BeginInvoke(action);
         }
 
         private void CancelButton_Click(object sender, EventArgs e)
@@ -52,9 +99,15 @@ namespace UpdateOp
             Close();
         }
 
-        private void SecondTimer_Tick(object sender, EventArgs e)
+        private void MainForm_FormClosing(object sender, FormClosingEventArgs e)
         {
-            ApplyUpdate();
+            KillThread = true;
+
+            if (Worker != null)
+            {
+                TryUpdate.Set();
+                Worker.Join();
+            }
         }
